[NEW] Update domains table on record create/update/delete [FIX] PM_DISABLE_SYSLOG renamed to DISABLE_SYSLOG [FIX] minor bug fixes

+++ public/install/tables/pm_domains.sql
@@ -8,9 +8,9 @@
 CREATE TABLE `pm_domains` (
   `rowid` int(11) NOT NULL AUTO_INCREMENT,
   `label` varchar(255) COLLATE table_collation NOT NULL,
-  `website` tinyint(1) DEFAULT 0,
-  `ftp` tinyint(1) DEFAULT 0,
-  `data_base` tinyint(1) DEFAULT 0,
+  `website` int(11) DEFAULT 0,
+  `ftp` int(11) DEFAULT 0,
+  `data_base` int(11) DEFAULT 0,
   `fk_user` int(11) NOT NULL,
   `tms` timestamp NOT NULL DEFAULT current_timestamp() ON UPDATE current_timestamp(),
   PRIMARY KEY (`rowid`),
